feat(csp): custom sign method
