+++ packages/react-providers/src/csp/index.ts
@@ -1,0 +1,26 @@
+import { CensusType, CspProofType, CspVote, PublishedElection, VocdoniSDKClient, Vote } from '@vocdoni/sdk'
+import { up } from 'up-fetch'
+
+const f = up(fetch)
+
+export const vote = async (client: VocdoniSDKClient, election: PublishedElection, token: string, vote: Vote) => {
+  if (!client) {
+    throw new Error('no client initialized')
+  }
+  if (!(election instanceof PublishedElection) || election?.census?.type !== CensusType.CSP) {
+    throw new Error('not a CSP election')
+  }
+
+  const walletAddress: string = (await client.wallet?.getAddress()) as string
+  const signature: string = await f(`${election.census.censusURI.replace(/\/bundle\/.+$/, '')}/${election.id}/sign`, {
+    method: 'POST',
+    body: {
+      payload: walletAddress,
+      tokenR: token,
+    },
+  })
+  const cspVote: CspVote = client.cspVote(vote, signature, CspProofType.ECDSA)
+  const vid: string = await client.submitVote(cspVote)
+
+  return vid
+}
